+++ dev/breeze/src/airflow_breeze/utils/docker_command_utils.py
@@ -58,6 +58,7 @@ from airflow_breeze.global_constants import (
     MIN_DOCKER_VERSION,
 )
 from airflow_breeze.utils.console import Output, console_print, get_console
+from airflow_breeze.utils.environment_check import check_uv_version
 from airflow_breeze.utils.run_utils import (
     RunCommandResult,
     check_if_buildx_plugin_installed,
@@ -372,82 +373,6 @@ Make sure docker-compose you install is first on the PATH variable of yours.\n
 [warning]At least {MIN_DOCKER_COMPOSE_VERSION} needed! Please upgrade!\n[/]
 See https://docs.docker.com/compose/install/ for installation instructions.\n
 Make sure docker-compose you install is first on the PATH variable of yours.\n
-"""
-        )
-        sys.exit(1)
-
-
-def _read_required_uv_version() -> str | None:
-    """Parse ``[tool.uv] required-version = ">=X.Y.Z"`` from the root ``pyproject.toml``.
-
-    Returns the raw ``X.Y.Z`` string, or ``None`` if the declaration is missing or
-    unparsable — in which case the caller should skip the check rather than fail the
-    environment bootstrap.
-    """
-    pyproject = (AIRFLOW_ROOT_PATH / "pyproject.toml").read_text()
-    # Narrow to the [tool.uv] section so we don't pick up an unrelated required-version.
-    section = re.search(r"^\[tool\.uv\]\s*$(?P<body>.*?)(?=^\[|\Z)", pyproject, re.MULTILINE | re.DOTALL)
-    if not section:
-        return None
-    match = re.search(
-        r"""^\s*required-version\s*=\s*["']\s*>=\s*(?P<ver>\d+(?:\.\d+){0,2})\s*["']""",
-        section.group("body"),
-        re.MULTILINE,
-    )
-    return match.group("ver") if match else None
-
-
-def check_uv_version(quiet: bool = False):
-    """Fail the environment check if ``uv`` on PATH is older than the project minimum.
-
-    The minimum is declared as ``[tool.uv] required-version`` in the root
-    ``pyproject.toml``. Matches the ``check_docker_version`` pattern: reports the
-    detected version on success and exits with install/upgrade guidance on failure.
-    """
-    required_uv_version = _read_required_uv_version()
-    if required_uv_version is None:
-        if not quiet:
-            console_print(
-                "[warning]Could not read `[tool.uv] required-version` from pyproject.toml — "
-                "skipping uv version check.[/]"
-            )
-        return
-    uv_version_result = run_command(
-        ["uv", "--version"],
-        no_output_dump_on_exception=True,
-        capture_output=True,
-        text=True,
-        check=False,
-        dry_run_override=False,
-    )
-    if uv_version_result.returncode != 0:
-        console_print(
-            f"""
-[error]`uv` is not installed or not on PATH.[/]
-[warning]Breeze needs uv >= {required_uv_version} to manage its environments.[/]
-See https://docs.astral.sh/uv/getting-started/installation/ for installation instructions.
-"""
-        )
-        sys.exit(1)
-    match = re.search(r"\b(\d+\.\d+(?:\.\d+)?)", uv_version_result.stdout)
-    if not match:
-        if not quiet:
-            console_print(
-                f"[warning]Unexpected `uv --version` output: {uv_version_result.stdout!r} — "
-                "skipping uv version check.[/]"
-            )
-        return
-    uv_version = match.group(1)
-    if compare_version(uv_version, required_uv_version):
-        if not quiet:
-            console_print(f"[success]Good version of uv: {uv_version}.[/]")
-    else:
-        console_print(
-            f"""
-[error]Your version of uv is too old: {uv_version}.\n[/]
-[warning]Please upgrade to at least {required_uv_version}.\n[/]
-Run `uv self update` (or reinstall following
-https://docs.astral.sh/uv/getting-started/installation/).
 """
         )
         sys.exit(1)
